feat - refactor /chat/completions to have a common helper
