+++ .pre-commit-config.yaml
@@ -1,14 +1,15 @@
 repos:
--   repo: https://github.com/astral-sh/ruff-pre-commit
-    # Ruff version.
-    rev: v0.4.8
+-   repo: https://github.com/psf/black
+    rev: 24.2.0
     hooks:
-        # Run the linter.
-        - id: ruff
-          exclude: ^litellm/tests/|^litellm/proxy/proxy_cli.py|^litellm/proxy/tests/
-        # Run the formatter.
-        - id: ruff-format
-          exclude: ^litellm/tests/|^litellm/proxy/proxy_cli.py|^litellm/proxy/tests/
+    - id: black
+-   repo: https://github.com/pycqa/flake8
+    rev: 7.0.0  # The version of flake8 to use
+    hooks:
+    -  id: flake8
+       exclude: ^litellm/tests/|^litellm/proxy/proxy_cli.py|^litellm/proxy/tests/
+       additional_dependencies: [flake8-print]
+       files: litellm/.*\.py
 -   repo: local
     hooks:
     -   id: check-files-match
